+++ geotrek/trekking/views.py
@@ -20,7 +20,7 @@ from geotrek.common.forms import AttachmentAccessibilityForm
 from geotrek.common.functions import Length
 from geotrek.common.mixins.api import APIViewSet
 from geotrek.common.mixins.forms import FormsetMixin
-from geotrek.common.mixins.views import CustomColumnsMixin, MetaMixin
+from geotrek.common.mixins.views import CompletenessMixin, CustomColumnsMixin, MetaMixin
 from geotrek.common.models import Attachment, RecordSource, TargetPortal, Label
 from geotrek.common.permissions import PublicOrReadPermMixin
 from geotrek.common.views import DocumentPublic, DocumentBookletPublic, MarkupPublic
@@ -32,6 +32,7 @@ from geotrek.infrastructure.serializers import InfrastructureAPIGeojsonSerialize
 from geotrek.signage.models import Signage
 from geotrek.signage.serializers import SignageAPIGeojsonSerializer
 from geotrek.zoning.models import District, City, RestrictedArea
+
 from .filters import TrekFilterSet, POIFilterSet, ServiceFilterSet
 from .forms import TrekForm, TrekRelationshipFormSet, POIForm, WebLinkCreateFormPopup, ServiceForm
 from .models import Trek, POI, WebLink, Service, TrekRelationship, OrderedTrekChild
@@ -104,7 +105,7 @@ class TrekKMLDetail(LastModifiedMixin, PublicOrReadPermMixin, BaseDetailView):
         return response
 
 
-class TrekDetail(MapEntityDetail):
+class TrekDetail(CompletenessMixin, MapEntityDetail):
     queryset = Trek.objects.existing()
 
     @property
@@ -335,7 +336,7 @@ class POIFormatList(MapEntityFormat, POIList):
             yield poi
 
 
-class POIDetail(MapEntityDetail):
+class POIDetail(CompletenessMixin, MapEntityDetail):
     queryset = POI.objects.existing()
 
     def get_context_data(self, *args, **kwargs):